Add Unit Test Modals with styling and functionality
- Implemented UnitTestModals component with multiple modal states (Unit Test, Model Selector, Choice, Queue Check).
- Added CSS styles for modals, including transitions and responsive design.
- Integrated keyboard handling for closing modals and sending messages.
- Created a queue system for managing unit test requests with unique tokens.
- Ensured proper rendering of modals using React portals and client-side checks.
